Began the werewolf
I began the werewolf, now it's a hero by itself, I made him a portrait
so it won't be obvious who changed to the werewolf.
Also, now there are some more variables to be added in the player class.

PS: Need to look in gameAPIs of the minimap for "MinimapDebugPoint"